pyproject.toml
[project]
name = "aivora"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Shymchuk",email = "paintool1@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.12.7"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
